Extract tests for NestNonPublicMethods rewriter
